# api/openapi.yaml
openapi: 3.1.0
info:
  title: Omahab Control API
  description: |
    Authenticated control-plane API for omahabd. All routes are under /api/v1 except
    /up. Bearer authentication is required except for /up and the email webhook which
    uses a separate HMAC authentication path. Mutating requests require Content-Type
    application/json. Unknown JSON fields are rejected. Request bodies are limited
    (1 MiB JSON, 5 MiB email webhook) and per-request timeouts are enforced. SSE
    streaming at /api/v1/events/stream supports Last-Event-ID replay and heartbeat
    comments. Secret and provider-credential reads return metadata only.
  version: 0.1.0
  contact:
    name: Omahab
servers:
  - url: http://localhost:8484
    description: omahabd default listen address
security:
  - bearerAuth: []
paths:
  /up:
    get:
      summary: Liveness probe
      description: Unauthenticated liveness check.
      operationId: getUp
      security: []
      responses:
        '200':
          description: Service is up
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status: { type: string, example: up }
                  version: { type: string }
        '500':
          $ref: '#/components/responses/Error'
  /api/v1/status:
    get:
      summary: Get control-plane status
      operationId: getStatus
      responses:
        '200':
          description: Status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Status' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/Error' }
  /api/v1/instance:
    get:
      summary: Get instance identity
      operationId: getInstance
      responses:
        '200':
          description: Instance
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Instance' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/Error' }
  /api/v1/catalog:
    get:
      summary: List installable curated bundles
      description: Runtime catalog shipped with signed releases. Empty when no release catalog is present; installs then fail closed.
      operationId: listCatalog
      responses:
        '200':
          description: Catalog bundles
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/CatalogBundle' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/Error' }
  /api/v1/applications:
    get:
      summary: List applications
      operationId: listApplications
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Applications
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Application' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/Error' }
    post:
      summary: Install a curated bundle
      description: Installs a catalog bundle by id with digest-pinned images. Non-private exposure requires a hostname.
      operationId: installApplication
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InstallApplicationRequest' }
      responses:
        '201': { description: Installed, content: { application/json: { schema: { $ref: '#/components/schemas/Application' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /api/v1/applications/{id}:
    get:
      summary: Get application
      operationId: getApplication
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200': { description: Application, content: { application/json: { schema: { $ref: '#/components/schemas/Application' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/Error' }
    patch:
      summary: Update application
      operationId: updateApplication
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateApplicationRequest' }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Application' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/Error' }
  /api/v1/applications/{id}/actions:
    post:
      summary: Perform application action
      description: Lifecycle actions (start, stop, restart, update, rollback). Installing a bundle is POST /api/v1/applications.
      operationId: doApplicationAction
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApplicationActionRequest' }
      responses:
        '200': { description: Result, content: { application/json: { schema: { $ref: '#/components/schemas/Application' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/Error' }
  /api/v1/applications/{id}/exposure:
    get:
      summary: Get application exposure
      operationId: getApplicationExposure
      parameters: [{ $ref: '#/components/parameters/IdPath' }]
      responses:
        '200': { description: Exposure, content: { application/json: { schema: { $ref: '#/components/schemas/ExposureState' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      summary: Update application exposure
      operationId: updateApplicationExposure
      parameters: [{ $ref: '#/components/parameters/IdPath' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateExposureRequest' } } } }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/ExposureState' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/exposure:
    get:
      summary: List all exposure states
      operationId: listExposure
      responses:
        '200':
          description: Exposure list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties: { items: { type: array, items: { $ref: '#/components/schemas/ExposureState' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/exposure/{kind}/{id}:
    parameters:
      - name: kind
        in: path
        required: true
        schema: { type: string, enum: [application, project] }
      - $ref: '#/components/parameters/IdPath'
    get:
      summary: Get exposure by kind
      operationId: getExposure
      responses:
        '200': { description: Exposure, content: { application/json: { schema: { $ref: '#/components/schemas/ExposureState' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      summary: Update exposure by kind
      operationId: updateExposure
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateExposureRequest' } } } }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/ExposureState' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/projects:
    get:
      summary: List projects
      operationId: listProjects
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Projects
          content: { application/json: { schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: '#/components/schemas/Project' } } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      summary: Create project
      operationId: createProject
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateProjectRequest' } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/projects/{id}:
    parameters: [{ $ref: '#/components/parameters/IdPath' }]
    get:
      summary: Get project
      operationId: getProject
      responses:
        '200': { description: Project, content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      summary: Update project
      operationId: updateProject
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateProjectRequest' } } } }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: Delete project
      operationId: deleteProject
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/projects/{id}/exposure:
    parameters: [{ $ref: '#/components/parameters/IdPath' }]
    get:
      summary: Get project exposure
      operationId: getProjectExposure
      responses:
        '200': { description: Exposure, content: { application/json: { schema: { $ref: '#/components/schemas/ExposureState' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      summary: Update project exposure
      operationId: updateProjectExposure
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateExposureRequest' } } } }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/ExposureState' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/v1/projects/{id}/releases:
    parameters: [{ $ref: '#/components/parameters/IdPath' }]
    get:
      summary: List releases
      operationId: listReleases
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200': { description: Releases, content: { application/json: { schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: '#/components/schemas/Release' } } } } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      summary: Create release
      operationId: createRelease
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateReleaseRequest' } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Release' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/v1/projects/{id}/releases/{releaseID}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
      - name: releaseID
        in: path
        required: true
        schema: { type: string }
    get:
      summary: Get release
      operationId: getRelease
      responses:
        '200': { description: Release, content: { application/json: { schema: { $ref: '#/components/schemas/Release' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/projects/{id}/releases/{releaseID}/rollback:
    post:
      summary: Rollback to release
      operationId: rollbackRelease
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: releaseID
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Rolled back, content: { application/json: { schema: { $ref: '#/components/schemas/Release' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/secrets:
    get:
      summary: List secrets (metadata only)
      description: Secret values are never returned. Each entry is metadata.
      operationId: listSecrets
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - name: scope
          in: query
          schema: { type: string }
          description: Filter by scope
      responses:
        '200': { description: Secrets, content: { application/json: { schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: '#/components/schemas/Secret' } } } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      summary: Create secret
      operationId: createSecret
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateSecretRequest' } } } }
      responses:
        '201': { description: Created metadata, content: { application/json: { schema: { $ref: '#/components/schemas/Secret' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/secrets/{id}:
    parameters: [{ $ref: '#/components/parameters/IdPath' }]
    get:
      summary: Get secret metadata
      operationId: getSecret
      responses:
        '200': { description: Metadata, content: { application/json: { schema: { $ref: '#/components/schemas/Secret' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      summary: Rotate secret value
      operationId: updateSecret
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateSecretRequest' } } } }
      responses:
        '200': { description: Updated metadata, content: { application/json: { schema: { $ref: '#/components/schemas/Secret' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: Delete secret
      operationId: deleteSecret
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/backups:
    get:
      summary: List backups
      operationId: listBackups
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200': { description: Backups, content: { application/json: { schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: '#/components/schemas/Backup' } } } } } } }
    post:
      summary: Trigger backup
      operationId: createBackup
      requestBody:
        required: false
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateBackupRequest' } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Backup' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/v1/backups/{id}:
    parameters: [{ $ref: '#/components/parameters/IdPath' }]
    get:
      summary: Get backup
      operationId: getBackup
      responses:
        '200': { description: Backup, content: { application/json: { schema: { $ref: '#/components/schemas/Backup' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/backups/{id}/restore:
    post:
      summary: Restore backup
      operationId: restoreBackup
      parameters: [{ $ref: '#/components/parameters/IdPath' }]
      responses:
        '200': { description: Restoring, content: { application/json: { schema: { $ref: '#/components/schemas/Backup' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/backups/{id}/verify:
    post:
      summary: Verify backup
      operationId: verifyBackup
      parameters: [{ $ref: '#/components/parameters/IdPath' }]
      responses:
        '200': { description: Verified, content: { application/json: { schema: { $ref: '#/components/schemas/Backup' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/events:
    get:
      summary: List events
      operationId: listEvents
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - name: type
          in: query
          schema: { type: string }
        - name: severity
          in: query
          schema: { type: string }
        - name: unread
          in: query
          schema: { type: boolean }
      responses:
        '200': { description: Events, content: { application/json: { schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: '#/components/schemas/Event' } } } } } } }
...
  /api/v1/events/{id}:
    parameters: [{ $ref: '#/components/parameters/IdPath' }]
    get:
      summary: Get event
      operationId: getEvent
      responses:
        '200': { description: Event, content: { application/json: { schema: { $ref: '#/components/schemas/Event' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Bearer token verified via constant-time SHA256 comparison. Required for all /api/v1 routes except /up and /api/v1/email/ingest.
    hmacAuth:
      type: apiKey
      in: header
      name: X-Omahab-Signature
      description: Worker v1 canonical HMAC-SHA256, optionally prefixed with sha256=. Metadata mirror headers are also required.
  parameters:
...
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0 }
...
  responses:
...
    Error:
      description: Error
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
  schemas:
...
    Status:
      type: object
      required: [instance_id, version, health, started_at, now]
      properties:
        instance_id: { type: string }
        version: { type: string }
        health: { $ref: '#/components/schemas/Health' }
        started_at: { type: string, format: date-time }
        now: { type: string, format: date-time }
...